split utils to api and impl
