+++ hydra/conf/__init__.py
@@ -24,12 +24,13 @@ hydra_defaults = [
 ]
 
 
-@dataclass
 # This extends Dict[str, Any] to allow for the deprecated "class" field.
 # Once support for class field removed this can stop extending Dict.
+@dataclass
 class PluginConf(Dict[str, Any]):
-    # class name for plugin
+    # class, class method or function name
     cls: str = MISSING
+    # parameters to pass to cls when calling it
     params: Any = field(default_factory=dict)
